#493 - Password Reset Mutation
- Pulled in the PR from @kellenmace and made changes to fit the post-v0.1.0 architecture
