+++ src/TypeRegistry.php
@@ -24,6 +24,7 @@ use WPGraphQL\Mutation\MediaItemUpdate;
 use WPGraphQL\Mutation\PostObjectCreate;
 use WPGraphQL\Mutation\PostObjectDelete;
 use WPGraphQL\Mutation\PostObjectUpdate;
+use WPGraphQL\Mutation\ResetUserPassword;
 use WPGraphQL\Mutation\TermObjectCreate;
 use WPGraphQL\Mutation\TermObjectDelete;
 use WPGraphQL\Mutation\TermObjectUpdate;
@@ -220,6 +221,7 @@ class TypeRegistry {
 		MediaItemCreate::register_mutation();
 		MediaItemDelete::register_mutation();
 		MediaItemUpdate::register_mutation();
+		ResetUserPassword::register_mutation();
 		UserCreate::register_mutation();
 		UserDelete::register_mutation();
 		UserUpdate::register_mutation();
